moved extension functions to where they are needed
