+++ game/communication/src/main/kotlin/de/gleex/pltcmd/game/communication/EventScopes.kt
@@ -1,0 +1,12 @@
+package de.gleex.pltcmd.game.communication
+
+import org.hexworks.cobalt.events.api.EventScope
+
+/**
+ * This scope is used for radio communications. It basically represents the radio network.
+ *
+ * Events used for this scope are [TransmissionEvent]s.
+ *
+ * @see [subscribeToRadioComms]
+ */
+internal object RadioComms : EventScope
