+++ game/ticks/src/main/kotlin/de/gleex/pltcmd/game/ticks/EventScopes.kt
@@ -1,0 +1,8 @@
+package de.gleex.pltcmd.game.ticks
+
+import org.hexworks.cobalt.events.api.EventScope
+
+/**
+ * Scope to publish game ticks and thus advance game time.
+ */
+internal object Ticks : EventScope
