+++ game/ticks/src/main/kotlin/de/gleex/pltcmd/game/ticks/TickEvent.kt
@@ -1,6 +1,9 @@
 package de.gleex.pltcmd.game.ticks
 
 import org.hexworks.cobalt.events.api.Event
+import org.hexworks.cobalt.events.api.EventBus
+import org.hexworks.cobalt.events.api.Subscription
+import org.hexworks.cobalt.events.api.simpleSubscribeTo
 
 /**
  * This event denotes that a "tick" happened. A tick has a unique ID and advances the game time which can be
@@ -10,4 +13,11 @@ class TickEvent(val id: TickId) : Event {
     override val emitter: Any
         get() = Ticker
 
+}
+
+/**
+ * Method to subscribe to [TickEvent]s
+ */
+fun EventBus.subscribeToTicks(onEvent: (TickEvent) -> Unit): Subscription {
+    return simpleSubscribeTo(Ticks, onEvent)
 }
